refactor(summer-school): split into separate big sections
- WhatIsItSection: just the title
- DetailsSection: first image + dates/location table
- PillarsSection: second image + Build/Reflect/Explore cards
